프로젝트 전체 리팩토링
- 디렉터리 구조 대 개편중
  -> scss 디렉터리 구조 수정중

- BanPickBoard 컴포넌트 하위 컴포넌트 분리 진행중.
  -> BanChampCard, GoalBoard, MatchResult, selectBoard.js 분리 진행중

- .scss 파일 나누기 및 className 수정중.
  -> 컴포넌트 분리에 따라 동시 진행.
