--- frontend/src/Layout/Components/index.js
@@ -1,3 +1,0 @@
-export { default as Footer } from './Footer'
-export { default as Header } from './Header'
-
