+++ frontend/src/components/BanpickBoard/GoalBoard/GoalViewer/index.js
@@ -1,0 +1,11 @@
+import React from 'react'
+import '@toast-ui/editor/dist/toastui-editor-viewer.css'
+import { Viewer } from '@toast-ui/react-editor'
+
+export default function GoalViewer() {
+  return (
+    <Viewer
+    // initialValue={viewerInput}
+    />
+  )
+}
