+++ frontend/src/components/BanpickBoard/index.js
@@ -2,15 +2,7 @@ import React, { useEffect, useState, useRef, useCallback } from 'react'
 import ReactTooltip from 'react-tooltip'
 import Hangul from 'hangul-js'
 import { Container, Row, Col } from 'react-bootstrap'
-import { Editor, Viewer } from '@toast-ui/react-editor'
-import { getDownloadResultPngFile } from '../../apis/get'
 
-import '@toast-ui/editor/dist/toastui-editor.css'
-import '@toast-ui/editor/dist/toastui-editor-viewer.css'
-import '@toast-ui/editor/dist/i18n/ko-kr'
-import 'tui-color-picker/dist/tui-color-picker.css'
-import '@toast-ui/editor-plugin-color-syntax/dist/toastui-editor-plugin-color-syntax.css'
-import colorSyntax from '@toast-ui/editor-plugin-color-syntax'
 import {
   searchIcon,
   transparencyImg,
@@ -20,6 +12,10 @@ import {
 import MatchInfo from './MatchInfo'
 import SummonerCard from './SummonerCard'
 import TeamSelectMenu from './TeamSelectMenu'
+import SelectBoard from './SelectBoard'
+import GoalBoard from './GoalBoard'
+import BanChampCard from './BanChampCard'
+import MatchResult from './MatchResult'
 
 export default function BanpickBoard({
   boardPhase,
@@ -28,30 +24,7 @@ export default function BanpickBoard({
   ascendingChampionDataList,
   classicSpellList,
 }) {
-  const resultDownToolTip1 = `밴픽 결과를 이미지 파일로 변환하여 자동 다운로드 합니다.<br>
-  주의하세요💩 : 소환사 챔피언 이미지와 일부 텍스트 사이즈가 조정되어 보이는 화면과 차이가 있습니다.<br>
-  (추후 사이즈 조정 없이 저장 가능하도록 업데이트 예정입니다^^)`
-  const resultDownToolTip2 = `결과 캡쳐를 위해 버튼 탭이 사라지고, 전체화면으로 전환되며, 스크롤이 조정됩니다.<br>  
-  (업데이트와 함께 사라질 기능입니다^^)`
-  const searchToolTip = `기본, 초성 검색이 가능합니다🧐<br>
-  띄어쓰기도 걱정하지 마세요! ex)리 신, 탐 켄치 
-  `
-
-  const teamArr = [
-    'KDF',
-    'T1',
-    'DK',
-    'BRO',
-    'DRX',
-    'GEN',
-    'HLE',
-    'KT',
-    'LSB',
-    'NS',
-  ]
-
   const isMountedRef = useRef(false)
-  const editorRef = useRef()
 
   const [globalPhase, setGlobalPhase] = useState('PickBan') // PickBan, GoalEdit, End
   const [pickBanPhase, setPickBanPhase] = useState('Pick') // Pick, Ban, Spell, End
@@ -60,33 +33,6 @@ export default function BanpickBoard({
   const [currentSelectingIndex, setCurrentSelectingIndex] = useState(0) // 0 ~ 4
   const [currentSelectingSpellNumber, setCurrentSelectingSpellNumber] =
     useState(1) // 1, 2
-
-  const [champDataList, setChampDataList] = useState([])
-  const [selectedBlueTeam, setSelectedBlueTeam] = useState('Blue')
-  const [selectedRedTeam, setSelectedRedTeam] = useState('Red')
-  const [isTeamSelectMenuOpen, setIsTeamSelectMenuOpen] = useState({
-    blue: false,
-    red: false,
-  })
-
-  const [searchInput, setSearchInput] = useState('')
-
-  const [matchResult, setMatchResult] = useState('Win or Lose')
-  const [goalPatchVersion, setGoalPatchVersion] = useState('Patch version : ')
-  const [viewerInput, setViewerInput] = useState('')
-
-  const [player, setPlayer] = useState({
-    blue1: '',
-    blue2: '',
-    blue3: '',
-    blue4: '',
-    blue5: '',
-    red1: '',
-    red2: '',
-    red3: '',
-    red4: '',
-    red5: '',
-  })
 
   class Summoner {
     constructor({
@@ -156,65 +102,6 @@ export default function BanpickBoard({
     new Summoner({}),
   ])
 
-  const onChangeSearchInput = (e) => setSearchInput(e.target.value)
-  const onChangePlayer = (e, teamNumber) =>
-    setPlayer({ ...player, [teamNumber]: e.target.value })
-  const onChangeGoalPatchVersion = (e) => setGoalPatchVersion(e.target.value)
-  const onChangeEditor = () => {
-    const editorInputHtml = editorRef.current.getInstance().getHTML()
-    setViewerInput(editorInputHtml)
-  }
-
-  const updateMatchChampDataList = useCallback(() => {
-    const championNameList = ascendingChampionDataList.map((data) => data.name)
-    let searcher = new Hangul.Searcher(searchInput)
-
-    const letterMatchedChampNameList = championNameList.filter(
-      (championName) => searcher.search(championName) >= 0
-    )
-    const letterMatchedChampNameListWithoutSpace = championNameList.filter(
-      (championName) => searcher.search(championName.replace(/ /gi, '')) >= 0
-    )
-    const chosungMatchedChampNameList = championNameList.filter(
-      (championName) => {
-        const champChosungStrArr = Hangul.d(championName, true)
-          .map((disEachLetterList) => disEachLetterList[0]) // ['ㄱ', 'ㄹ'] and ['ㄱ', 'ㄹ', 'ㅇ'] ...something
-          .join('') // ['ㄱㄹ'] and ['ㄱㄹㅇ'] ...something
-          .replace(/ /gi, '') // 띄어쓰기 제거
-
-        const searchInputChosungStrArr = Hangul.d(searchInput).join('') // ['ㄱ','ㄹ'] -> ['ㄱㄹ']
-        return champChosungStrArr.includes(searchInputChosungStrArr)
-      }
-    )
-
-    const mergedChampNameList = letterMatchedChampNameList.concat(
-      letterMatchedChampNameListWithoutSpace,
-      chosungMatchedChampNameList
-    )
-    const duplicatesRemovedChampNameList = mergedChampNameList.filter(
-      (name, index) => mergedChampNameList.indexOf(name) === index
-    )
-
-    const matchedChampDataList = ascendingChampionDataList.filter((data) =>
-      duplicatesRemovedChampNameList.includes(data.name)
-    )
-
-    setChampDataList(matchedChampDataList)
-  }, [ascendingChampionDataList, searchInput])
-
-  const toggleIsTeamSelectMenuOpen = (teamColor) => {
-    setIsTeamSelectMenuOpen((prevState) => ({
-      ...prevState,
-      [teamColor]: !prevState[teamColor],
-    }))
-  }
-  const closeTeamSelectMenu = (teamColor) => {
-    setIsTeamSelectMenuOpen((prevState) => ({
-      ...prevState,
-      [teamColor]: false,
-    }))
-  }
-
   const updateSummonerData = ({ type, data, isConfirmed }) => {
     let [team, setTeam] =
       currentSelectingTeam === 'blue'
@@ -263,17 +150,17 @@ export default function BanpickBoard({
     return team[currentSelectingIndex].isEmpty(type)
   }
 
-  const onClickChampionPickButton = () => {
-    if (!isCurrentSelectingDataEmpty(currentSelectingType())) {
-      updateSummonerData({ type: 'pickedChampion', isConfirmed: true })
-    }
-  }
+  // const onClickChampionPickButton = () => {
+  //   if (!isCurrentSelectingDataEmpty(currentSelectingType())) {
+  //     updateSummonerData({ type: 'pickedChampion', isConfirmed: true })
+  //   }
+  // }
 
-  const onClickChampionBanButton = () => {
-    if (!isCurrentSelectingDataEmpty(currentSelectingType())) {
-      updateSummonerData({ type: 'bannedChampion', isConfirmed: true })
-    }
-  }
+  // const onClickChampionBanButton = () => {
+  //   if (!isCurrentSelectingDataEmpty(currentSelectingType())) {
+  //     updateSummonerData({ type: 'bannedChampion', isConfirmed: true })
+  //   }
+  // }
 
   const onClickSpellSelectButton = () => {
     if (
@@ -291,22 +178,6 @@ export default function BanpickBoard({
       setTeam(updatedArr)
     }
   }
-
-  const activatePlayerPrefix = useCallback(() => {
-    setPlayer((prev) => ({
-      ...prev,
-      blue1: `${selectedBlueTeam} TOP`,
-      blue2: `${selectedBlueTeam} JGL`,
-      blue3: `${selectedBlueTeam} MID`,
-      blue4: `${selectedBlueTeam} BOT`,
-      blue5: `${selectedBlueTeam} SUP`,
-      red1: `${selectedRedTeam} TOP`,
-      red2: `${selectedRedTeam} JGL`,
-      red3: `${selectedRedTeam} MID`,
-      red4: `${selectedRedTeam} BOT`,
-      red5: `${selectedRedTeam} SUP`,
-    }))
-  }, [selectedBlueTeam, selectedRedTeam])
 
   const activateSpellPlaceholder = useCallback(() => {
     const blueTeamUpdateArr = blueTeamSummoner.map((summoner) =>
@@ -623,12 +494,6 @@ export default function BanpickBoard({
   }, [blueTeamSummoner, redTeamSummoner])
 
   useEffect(() => {
-    //activatePlayerPrefix
-    activatePlayerPrefix()
-    console.log('activate PlayerPrefix')
-  }, [activatePlayerPrefix])
-
-  useEffect(() => {
     //activateSpellPlaceholder
     if (mode === 'rapid' && isMountedRef.current === true) {
       activateSpellPlaceholder()
@@ -639,373 +504,52 @@ export default function BanpickBoard({
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [boardPhase])
 
-  useEffect(() => {
-    // updateMatchChampDataList
-    updateMatchChampDataList()
-    console.log('matchedChampDataList update compleat')
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [searchInput, boardPhase])
+  // useEffect(() => {
+  //   setSearchInput('')
+  //   console.log('Initialize search input')
+  // }, [currentSelectingIndex])
 
-  useEffect(() => {
-    setSearchInput('')
-    console.log('Initialize search input')
-  }, [currentSelectingIndex])
+  const [blueTeamName, setBlueTeamName] = useState('Blue')
+  const [redTeamName, setRedTeamName] = useState('Red')
+
+  const redTeamInlineStyle = {
+    teamSelectMenu: { flexDirection: 'row-reverse' },
+    nameSelect: { right: 0, left: `${20}%` },
+  }
 
   return (
     <Container id="ban-pick-board" className="ban-pick-board">
       <Row id="board-top" className="board-top">
-        <TeamSelectMenu />
+        <TeamSelectMenu
+          teamColor={'blue'}
+          teamName={blueTeamName}
+          setTeamName={setBlueTeamName}
+          inlineStyle={{}}
+        />
         <MatchInfo />
-        <TeamSelectMenu />
+        <TeamSelectMenu
+          teamColor={'red'}
+          teamName={redTeamName}
+          setTeamName={setRedTeamName}
+          inlineStyle={redTeamInlineStyle}
+        />
       </Row>
 
       <Row className="board-middle">
         <div className="blue-team__summoners">
           {blueTeamSummoner.map((summoner, index) => (
-            <SummonerCard />
+            <SummonerCard key={index} />
           ))}
         </div>
 
         {globalPhase === 'PickBan' && (
-          <>
-            {pickBanPhase === 'Pick' && (
-              <div className="champion-select-board">
-                <div className="champion__select-option">
-                  <div className="search">
-                    <label className="search-input-label" htmlFor="search">
-                      <img
-                        className="search-icon"
-                        src={searchIcon}
-                        alt="search-icon"
-                        data-for="search-tooltip"
-                        data-tip={searchToolTip}
-                      />
-                      <ReactTooltip
-                        id="search-tooltip"
-                        multiline={true}
-                        delayShow={100}
-                      />
-                    </label>
-                    <input
-                      id="search"
-                      className="search-input"
-                      type="text"
-                      placeholder="챔피언 이름 검색"
-                      spellCheck="false"
-                      value={searchInput}
-                      onChange={(e) => onChangeSearchInput(e)}
-                    />
-                  </div>
-                </div>
-                <div className="champions">
-                  {champDataList.map((championData, index) => (
-                    <div
-                      className="champion__card"
-                      key={index}
-                      data-champion={championData.id}
-                      data-picked={isPickedChampion(championData.id)}
-                      data-banned={isBannedChampion(championData.id)}
-                      onClick={() => {
-                        updateSummonerData({
-                          type: 'pickedChampion',
-                          data: championData.id,
-                          isConfirmed: false,
-                        })
-                      }}
-                    >
-                      <img
-                        className="champion__img"
-                        alt={championData.id}
-                        src={`${process.env.REACT_APP_API_BASE_URL}/cdn/${recentVersion}/img/champion/${championData.id}.png`}
-                      />
-                      <small className="champion__name">
-                        {championData.name}
-                      </small>
-                    </div>
-                  ))}
-                </div>
-                <input
-                  className="champion__select-button"
-                  type="button"
-                  value={pickBanPhase}
-                  onClick={() => {
-                    onClickChampionPickButton()
-                  }}
-                />
-              </div>
-            )}
-
-            {pickBanPhase === 'Ban' && (
-              <div className="champion-select-board">
-                <div className="champion__select-option">
-                  <div className="search">
-                    <label className="search-input-label" htmlFor="search">
-                      <img
-                        className="search-icon"
-                        src={searchIcon}
-                        alt="search-icon"
-                        data-for="search-tooltip"
-                        data-tip={searchToolTip}
-                      />
-                      <ReactTooltip
-                        id="search-tooltip"
-                        multiline={true}
-                        delayShow={100}
-                      />
-                    </label>
-                    <input
-                      id="search"
-                      className="search-input"
-                      type="text"
-                      placeholder="챔피언 이름 검색"
-                      spellCheck="false"
-                      value={searchInput}
-                      onChange={(e) => onChangeSearchInput(e)}
-                    />
-                  </div>
-                </div>
-                <div className="champions">
-                  <div
-                    className="champion__card"
-                    onClick={() => {
-                      updateSummonerData({
-                        type: 'bannedChampion',
-                        data: 'noBan',
-                        isConfirmed: false,
-                      })
-                    }}
-                  >
-                    <img
-                      className="champion__img"
-                      alt="no-ban-icon"
-                      src={noBanIcon}
-                    />
-                    <small className="champion__name">없음</small>
-                  </div>
-                  {champDataList.map((championData, index) => (
-                    <div
-                      className="champion__card"
-                      key={index}
-                      data-champion={championData.id}
-                      data-picked={isPickedChampion(championData.id)}
-                      data-banned={isBannedChampion(championData.id)}
-                      onClick={() => {
-                        updateSummonerData({
-                          type: 'bannedChampion',
-                          data: championData.id,
-                          isConfirmed: false,
-                        })
-                      }}
-                    >
-                      <img
-                        className="champion__img"
-                        alt={championData.id}
-                        src={`${process.env.REACT_APP_API_BASE_URL}/cdn/${recentVersion}/img/champion/${championData.id}.png`}
-                      />
-                      <small className="champion__name">
-                        {championData.name}
-                      </small>
-                    </div>
-                  ))}
-                </div>
-                <input
-                  className="champion__select-button"
-                  type="button"
-                  value={pickBanPhase}
-                  onClick={() => {
-                    onClickChampionBanButton()
-                  }}
-                />
-              </div>
-            )}
-
-            {pickBanPhase === 'Spell' && (
-              <div
-                className="spell-select-board"
-                data-helper-text-color={`${currentSelectingTeam}`}
-              >
-                <div className="spell__select-helper">
-                  <div className="select-helper__text">
-                    <p>
-                      {currentSelectingTeam === 'blue'
-                        ? `${player[`blue${currentSelectingIndex + 1}`]}`
-                        : `${player[`red${currentSelectingIndex + 1}`]}`}{' '}
-                      스펠 선택중입니다.
-                    </p>
-                  </div>
-                  <div className="select-helper__zoom-view">
-                    <img
-                      alt={`zoom-view-spell1`}
-                      data-current-target={currentSelectingSpellNumber === 1}
-                      onClick={() => {
-                        setCurrentSelectingSpellNumber(1)
-                      }}
-                      src={zoomViewImgSrc(1)}
-                    />
-                    <img
-                      alt={`zoom-view-spell2`}
-                      data-current-target={currentSelectingSpellNumber === 2}
-                      onClick={() => {
-                        setCurrentSelectingSpellNumber(2)
-                      }}
-                      src={zoomViewImgSrc(2)}
-                    />
-                  </div>
-                </div>
-                <div className="spells">
-                  {classicSpellList.map((spell, index) => (
-                    <div
-                      className="spell__card"
-                      key={index}
-                      data-picked-spell={isPickedSpell(spell.id)}
-                      onClick={() => {
-                        updateSpell(spell.id)
-                      }}
-                    >
-                      <img
-                        className="spell__img"
-                        alt={`spell-img-${spell.id}`}
-                        src={`${process.env.REACT_APP_API_BASE_URL}/cdn/${recentVersion}/img/spell/${spell.id}.png`}
-                      />
-                    </div>
-                  ))}
-                </div>
-                <div className="spell__select-button-wrap">
-                  <input
-                    className="spell__select-button"
-                    type="button"
-                    value={pickBanPhase}
-                    onClick={() => {
-                      onClickSpellSelectButton()
-                    }}
-                  />
-                </div>
-              </div>
-            )}
-          </>
+          <SelectBoard classicSpellList={classicSpellList} />
         )}
-
-        {globalPhase === 'GoalEdit' && (
-          <div id="todays-goal" className="todays-goal">
-            <div className="goal__patch-version-wrap">
-              <input
-                id="goal__patch-version"
-                className="goal__patch-version"
-                type="text"
-                value={goalPatchVersion}
-                onChange={(e) => onChangeGoalPatchVersion(e)}
-              />
-            </div>
-            {goalEditPhase === 'Editing' && (
-              <>
-                <div
-                  className="goal__editor-wrap"
-                  onClick={() => setGoalEditPhase('Editing')}
-                >
-                  <Editor
-                    initialValue={viewerInput}
-                    previewStyle="tap"
-                    hideModeSwitch={true}
-                    viewer={false}
-                    height="100%"
-                    minHeight="300px"
-                    initialEditType="wysiwyg"
-                    useCommandShortcut={false}
-                    language="ko-KR"
-                    ref={editorRef}
-                    onChange={() => {
-                      onChangeEditor()
-                    }}
-                    usageStatistics={false}
-                    plugins={[colorSyntax]}
-                    toolbarItems={[
-                      // 툴바 옵션 설정
-                      ['heading', 'bold', 'italic', 'strike'],
-                      ['hr', 'quote'],
-                      ['ul', 'ol'],
-                    ]}
-                  />
-                </div>
-                <div
-                  id="goal__button-wrap"
-                  className="goal__button-wrap"
-                  data-html2canvas-ignore
-                >
-                  <button onClick={() => setGoalEditPhase('EditDone')}>
-                    작성 완료
-                  </button>
-                </div>
-              </>
-            )}
-
-            {goalEditPhase === 'EditDone' && (
-              <>
-                <div
-                  className="goal__editor-wrap"
-                  onClick={() => setGoalEditPhase('Editing')}
-                >
-                  <Viewer initialValue={viewerInput} />
-                </div>
-                <div
-                  id="goal__button-wrap"
-                  className="goal__button-wrap"
-                  data-html2canvas-ignore
-                >
-                  <button
-                    data-for="button-tooltip1"
-                    data-tip={resultDownToolTip1}
-                    data-class="result-down-tooltip"
-                    onClick={() => getDownloadResultPngFile('ban-pick-board')}
-                  >
-                    결과 다운로드
-                  </button>
-                  <ReactTooltip
-                    id="button-tooltip1"
-                    multiline={true}
-                    delayShow={100}
-                  />
-
-                  <button
-                    data-for="button-tooltip2"
-                    data-tip={resultDownToolTip2}
-                    data-class="result-down-tooltip"
-                    onClick={() => {
-                      setGoalEditPhase('End')
-                      document.documentElement
-                        .requestFullscreen()
-                        .then(() =>
-                          document.documentElement.scroll(
-                            0,
-                            document.documentElement.clientHeight * 0.013
-                          )
-                        )
-                    }}
-                  >
-                    직접 캡쳐
-                  </button>
-                  <ReactTooltip
-                    id="button-tooltip2"
-                    multiline={true}
-                    delayShow={100}
-                  />
-                </div>
-              </>
-            )}
-
-            {goalEditPhase === 'End' && (
-              <div
-                className="goal__editor-wrap"
-                onClick={() => setGoalEditPhase('Editing')}
-              >
-                <Viewer initialValue={viewerInput} />
-              </div>
-            )}
-          </div>
-        )}
+        {globalPhase === 'GoalEdit' && <GoalBoard />}
 
         <div className="red-team__summoners">
           {redTeamSummoner.map((summoner, index) => (
-            <SummonerCard />
+            <SummonerCard key={index} />
           ))}
         </div>
       </Row>
@@ -1013,74 +557,15 @@ export default function BanpickBoard({
       <Row className="board-bottom">
         <div className="blue-team__ban">
           {blueTeamSummoner.map((summoner, index) => (
-            <div
-              className="banned-champion-wrap"
-              key={index}
-              data-current-target={
-                currentSelectingTeam === 'blue' &&
-                currentSelectingIndex === index &&
-                pickBanPhase === 'Ban'
-              }
-            >
-              <img
-                className="banned-champion"
-                alt={`blueTeam-banned-${index}-${summoner.bannedChampion.data}`}
-                data-current-target={
-                  currentSelectingTeam === 'blue' &&
-                  currentSelectingIndex === index &&
-                  pickBanPhase === 'Ban'
-                }
-                onClick={() => {
-                  setCurrentSelectingTeam('blue')
-                  setCurrentSelectingIndex(index)
-                  setPickBanPhase('Ban')
-                  setGlobalPhase('PickBan')
-                }}
-                src={bannedChampionImgSrc(summoner)}
-              />
-            </div>
+            <BanChampCard />
           ))}
         </div>
 
-        <div id="match-result-wrap" className="match-result-wrap">
-          <input
-            type="text"
-            id="match-result"
-            className="match-result"
-            value={matchResult}
-            spellCheck="false"
-            onChange={(e) => setMatchResult(e.target.value)}
-          />
-        </div>
+        <MatchResult />
 
         <div className="red-team__ban">
           {redTeamSummoner.map((summoner, index) => (
-            <div
-              className="banned-champion-wrap"
-              key={index}
-              data-current-target={
-                currentSelectingTeam === 'red' &&
-                currentSelectingIndex === index &&
-                pickBanPhase === 'Ban'
-              }
-            >
-              <img
-                className="banned-champion"
-                alt={`redTeam-banned-${index}-${summoner.bannedChampion.data}`}
-                data-current-target={
-                  currentSelectingTeam === 'red' &&
-                  currentSelectingIndex === index &&
-                  pickBanPhase === 'Ban'
-                }
-                onClick={() => {
-                  setCurrentSelectingTeam('red')
-                  setCurrentSelectingIndex(index)
-                  setPickBanPhase('Ban')
-                  setGlobalPhase('PickBan')
-                }}
-                src={bannedChampionImgSrc(summoner)}
-              />
-            </div>
+            <BanChampCard />
           ))}
         </div>
       </Row>
