+++ frontend/src/scss/Board/BanpickBoard/MatchInfo/_MatchInfo.scss
