+++ frontend/src/scss/Board/BanpickBoard/MatchResult/_MatchResult.scss
@@ -1,0 +1,20 @@
+.match-result-wrap {
+  display: flex;
+  flex: 1 1 20%;
+  margin-top: 20px;
+  justify-content: center;
+  position: relative;
+
+  .match-result {
+    border: none;
+    background-color: transparent;
+    color: #fff;
+    text-align: center;
+    font-size: 2vw;
+    font-weight: 700;
+    line-height: 1.5;
+    &:hover {
+      background-color: #bcbcbc78;
+    }
+  }
+}
