+++ frontend/src/scss/Board/BanpickBoard/SelectBoard/_SelectBoard.scss
@@ -1,0 +1,2 @@
+@import "./ChampSelectBoard/_ChampSelectBoard";
+@import "./SpellSelectBoard/_SpellSelectBoard";
