+++ frontend/src/scss/Board/BanpickBoard/SummonerCard/SummonerChamp/_SummonerChamp.scss
