+++ frontend/src/scss/Board/BanpickBoard/SummonerCard/_SummonerCard.scss
@@ -1,6 +1,6 @@
-@import "./SummonerChamp/SummonerChamp";
-@import "./SummonerName/SummonerName";
-@import "./SummonerSpell/SummonerSpell";
+@import "./SummonerChamp/_SummonerChamp";
+@import "./SummonerName/_SummonerName";
+@import "./SummonerSpell/_SummonerSpell";
 
 .summoner__card {
   position: relative;
