+++ frontend/src/scss/Board/BanpickBoard/TeamSelectMenu/_TeamSelectMenu.scss
@@ -71,60 +71,3 @@
     }
   }
 }
-
-// .red-team {
-//   cursor: pointer;
-
-//   .red-team__name {
-//     position: relative;
-//     display: flex;
-//     align-items: center;
-//     justify-content: center;
-//     text-align: center;
-//     min-width: 25%;
-//     height: 100%;
-
-//     .name__button {
-//       color: #fff;
-//       text-align: center;
-//       font-size: 3vw;
-//       line-height: 1.5;
-//       background-color: transparent;
-//       border: none;
-//     }
-//     .name__select {
-//       position: absolute;
-//       right: 20%;
-//       top: 100%;
-//       color: #fff;
-//       margin: 0;
-//       padding: 0;
-//       width: 100%;
-//       background-color: rgba(0, 0, 0, 0.8);
-
-//       .name__option {
-//         display: flex;
-//         align-items: center;
-//         border: solid 1px rgba(187, 187, 187, 0.151);
-//         width: 100%;
-//         &:hover {
-//           background-color: #bcbcbc78;
-//         }
-
-//         .option__logo {
-//           width: 40%;
-//           padding: 0.5vw;
-//         }
-
-//         .option__span {
-//           flex: 1 0;
-//           font-size: 1.5vw;
-//         }
-//       }
-//     }
-//   }
-
-//   .red-team__logo {
-//     height: 100%;
-//   }
-// }
