+++ frontend/src/scss/Board/_Board.scss
@@ -1,0 +1,2 @@
+@import "./BanpickBoard/_BanpickBoard";
+@import "./SettingBoard/_SettingBoard";
