+++ frontend/src/scss/Layout/Footer/_Footer.scss
