+++ frontend/src/scss/Layout/_Layout.scss
@@ -1,0 +1,2 @@
+@import "./Header/Header";
+@import "./Footer/Footer";
